Grafo/Query/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0

